add: StatusGroup component activities/prisma-schema#161
